Convert control character sequence name to actual character sequence

Convert control character sequence name to actual character sequence

as_line_feed_chr(eol = c("LF", "CRLF", "CR", "LFCR"))

Arguments

eol
End of line (EOL) control character sequence. One of

"LF" for the line feed (LF) character ("\n"). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the default.

"CRLF" for the carriage return + line feed (CR+LF) character sequence ("\r\n"). The standard on Microsoft Windows, DOS and some other systems.

"CR" for the carriage return (CR) character ("\r"). The standard on classic Mac OS and some other antiquated systems.

"LFCR" for the line feed + carriage return (LF+CR) character sequence ("\n\r"). The standard on RISC OS and some other exotic systems.

Value

A character scalar.

See also