List values as a Markdown list

Generates a Markdown list of the given values formatted as verbatim. Character values are additionally wrapped in double quotes.

Usage

as_md_val_list(...)

Arguments

...

Values to be formatted. One or more R objects.

Value

A character scalar.

See also

Examples

list(1L, 2.2, "other") |>
  pal::as_md_val_list() |>
  cat()
#> * `1` 
#> * `2.2` 
#> * `"other"` 
  
# note that values are flattened before listing them, so this yields the same list
list(list(list(1L, list(2.2), list(list("other"))))) |>
  pal::as_md_val_list() |>
  cat()
#> * `1` 
#> * `2.2` 
#> * `"other"`