Format values as verbatim Markdown

Converts the given values to a character vector, formatted as verbatim Markdown. Character values are additionally wrapped in double quotes.

as_md_vals(...)

Arguments

...

R objects to list.

Value

A character vector.

See also

Examples

list(1L, 2.2, "other") |>
  pal::as_md_vals()
#> [1] "`1`"         "`2.2`"       "`\"other\"`"
  
# note that values are flattened before further processing them, so this yields the same result
list(list(list(1L, list(2.2), list(list("other"))))) |>
  pal::as_md_vals()
#> [1] "`1`"         "`2.2`"       "`\"other\"`"