Assert an object is member of any of the specified classes

Assert an object is member of any of the specified classes

assert_class_any(x, classes, name = "x")

Arguments

x
R object to test.

classes
Class names to check for inheritance. A character vector.

name
Name of the checked object to print in error message in case the assertion fails. A character scalar.

Value

x, invisibly.

Examples

pal::gh_text_file(path = "README.md",
                  owner = "salim-b",
                  name = "pal") |>
  pal::md_to_xml() |>
  assert_class_any(classes = c("xml_document", "xml_nodeset", "xml_node"))