Assert a package is installed

Assert a package is installed

assert_pkg(pkg, message = NULL)

Arguments

pkg
Package name. A character scalar.

message
The error message to display in case the package is not installed. If NULL, defaults to a sensible standard message.

Value

The package name invisibly.

See also

Examples

assert_pkg("pal")

assert_pkg(pkg = "glue",
           message = paste0("You should really consider to install the awesome `glue` package! ",
                            "It's the glue that keeps strings and variables together 🤲."))