Capitalize first letter

Converts the input to a character vector, with the first letter of each element uppercased.

Usage

capitalize_first(x)

Arguments

x

Input of which to capitalize the first letter, typically a character vector.

Value

A character vector of the same length as x.

See also

Examples

pal::capitalize_first(c("one", "Two", "tHREE"))
#> [1] "One"   "Two"   "THREE"