Deploy pkgdown site to local Git folder

Copies the static pkgdown website files to another local Git folder, then stages, commits and pushes the changes. Use pkgdown::build_site() before running this function in order to create the website files.

Use this function with caution since by default it completely wipes the to_path directory!

deploy_pkgdown_site(
  pkg_path = ".",
  to_path = NULL,
  use_dev_build = NULL,
  clean_to_path = TRUE,
  commit_msg = paste0("auto-deploy pkgdown site for ", desc_value(file = pkg_path, key
    = "Package"))
)

Arguments

pkg_path
Path to the R package of which the pkgdown website files are to be deployed.

to_path
Path to the Git (sub)folder to which the pkgdown website files are to be deployed. If NULL, the path is looked up in the R option pal.local_pkgdown_deploy_paths, which must be a named character vector of the form c(pkg_name = "/path/to/local_git_repo").

use_dev_build
Whether to deploy the development build of the pkgdown website files.

clean_to_path
Whether to completely wipe to_path before deploying the new pkgdown website files. Setting this TRUE ensures there are no obsolete files left over from previous deployments.

commit_msg
The Git commit message used for the deployment.

Value

A vector of paths to the deployed files/folders, invisibly.

Details

Instead of directly providing a valid to_path argument, you can also set to_path = NULL and provide the path in the R option pal.local_pkgdown_deploy_paths. The value of this option must be a named character vector whereas the names correspond to the package names and the values to the filesystem paths. As an example:

options(pal.local_pkgdown_deploy_paths = c(some_pkg    = "/path/to/common_pkgdown_repo/some_pkg",
                                           another_pkg = "/path/to/common_pkgdown_repo/another_pkg",
                                           foo_pkg     = "/path/to/dedicated_foo_pkgdown_repo"))

A suitable place to define the above option is the .Rprofile file in the user's home directory.