Get value from DESCRIPTION file field, cleaned up and with dynamic fallback

Returns the value from a DESCRIPTION file field (aka key). Thin wrapper around desc::desc_get_field() that returns an informative string referring to the given key as default if the field is unset.

Usage

desc_get_field_safe(
  key,
  default = glue::glue("<No `{key}` field set in DESCRIPTION!>"),
  trim_ws = TRUE,
  squish_ws = trim_ws,
  file = "."
)

Arguments

key

The field to query.

default

Default value to return if key is not found.

trim_ws

Whether to trim leading and trailing whitespace from the value. Defaults to TRUE.

squish_ws

Whether to reduce repeated whitespace in the value. Defaults to trim_ws.

file

DESCRIPTION file to use. By default the DESCRIPTION file of the current package (i.e. the package the working directory is part of) is used.

Value

A character scalar.

Details

By default, the following string is returned if key = "NoRealKey" is not found:

"<No `NoRealKey` field set in DESCRIPTION!>"

If you rather want to take an action like throwing an error, it's recommended to call desc::desc_get_field() directly.

See also

Examples

pal::desc_get_field_safe(key = "Description",
                         file = fs::path_package("pal"))
#> [1] "A collection of convenience/utility functions complementing base R and the tidyverse."