Escape line feeds / newlines

Escapes the POSIX-standard newline control character LF (aka \n) which is the standard on Unix/Linux and recent versions of macOS. Set escape_cr = TRUE in order to also escape the carriage return character CR (aka \r) commonly used on Microsoft Windows.

escape_lf(x, escape_cr = FALSE)

Arguments

x
A character vector.

escape_cr
Whether or not to also escape the carriage return character CR (aka \r). A logical scalar.

Value

A character vector of the same length as x.

See also

Examples

# read in and print package description as-is
text <-
  fs::path_package(package = "pal",
                   "DESCRIPTION") %>%
  readr::read_file() %T>%
  cat_lines()

#> Package: pal
#> Title: Friendly Convenience/Utility Functions
#> Version: 0.0.9050
#> Authors@R: 
#>     person(given = "Salim",
#>            family = "Brüggemann",
#>            role = c("aut", "cre"),
#>            email = "salim-b@pm.me",
#>            comment = c(ORCID = "0000-0002-5329-5987"))
#> Description: A collection of convenience/utility functions which
#>     aren't (yet) found in any other tidyverse-style R package.
#> License: AGPL (>= 3)
#> URL: https://rpkg.dev/pal, https://gitlab.com/salim_b/r/pkgs/pal
#> BugReports: https://gitlab.com/salim_b/r/pkgs/pal/issues
#> Depends: R (>= 4.0)
#> Imports: checkmate (>= 2.0.0), dplyr (>= 1.0.2), fs (>= 1.4.1), glue
#>         (>= 1.4.1), lifecycle (>= 0.2.0), magrittr (>= 1.5), methods,
#>         purrr (>= 0.3.4), rlang (>= 0.4.4), stringr (>= 1.4.0), tibble
#>         (>= 2.1.3), utils
#> Suggests: brio (>= 1.1.0), cli (>= 2.0.2), commonmark (>= 1.7), crayon
#>         (>= 1.3.4), daff (>= 0.3.5), desc (>= 1.2.0), gert, httr (>=
#>         1.4.1), janeaustenr, knitr (>= 1.28), pander (>= 0.6.3),
#>         pkgdown, pkgsnip, readr (>= 1.4.0), rex, rmarkdown (>= 2.1),
#>         roxygen2 (>= 7.1.1), rprojroot (>= 1.3.2), rstudioapi (>=
#>         0.11), testthat (>= 3.0.0), waldo (>= 0.2.3), withr (>= 2.3.0),
#>         writexl (>= 1.3.1), xfun, xopen (>= 1.0.0), yesno (>= 0.1.2)
#> Remotes: github::r-lib/rprojroot#79, gitlab::salim_b/r/pkgs/pkgsnip
#> Config/testthat/edition: 3
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.1.1
#> Author: Salim Brüggemann [aut, cre] (<https://orcid.org/0000-0002-5329-5987>)
#> Maintainer: Salim Brüggemann <salim-b@pm.me>
#> Built: R 4.0.4; ; 2021-02-28 14:41:28 UTC; unix
#> 

# escape newlines and print again
escape_lf(text) %>% cat_lines()

#> Package: pal\nTitle: Friendly Convenience/Utility Functions\nVersion: 0.0.9050\nAuthors@R: \n    person(given = "Salim",\n           family = "Brüggemann",\n           role = c("aut", "cre"),\n           email = "salim-b@pm.me",\n           comment = c(ORCID = "0000-0002-5329-5987"))\nDescription: A collection of convenience/utility functions which\n    aren't (yet) found in any other tidyverse-style R package.\nLicense: AGPL (>= 3)\nURL: https://rpkg.dev/pal, https://gitlab.com/salim_b/r/pkgs/pal\nBugReports: https://gitlab.com/salim_b/r/pkgs/pal/issues\nDepends: R (>= 4.0)\nImports: checkmate (>= 2.0.0), dplyr (>= 1.0.2), fs (>= 1.4.1), glue\n        (>= 1.4.1), lifecycle (>= 0.2.0), magrittr (>= 1.5), methods,\n        purrr (>= 0.3.4), rlang (>= 0.4.4), stringr (>= 1.4.0), tibble\n        (>= 2.1.3), utils\nSuggests: brio (>= 1.1.0), cli (>= 2.0.2), commonmark (>= 1.7), crayon\n        (>= 1.3.4), daff (>= 0.3.5), desc (>= 1.2.0), gert, httr (>=\n        1.4.1), janeaustenr, knitr (>= 1.28), pander (>= 0.6.3),\n        pkgdown, pkgsnip, readr (>= 1.4.0), rex, rmarkdown (>= 2.1),\n        roxygen2 (>= 7.1.1), rprojroot (>= 1.3.2), rstudioapi (>=\n        0.11), testthat (>= 3.0.0), waldo (>= 0.2.3), withr (>= 2.3.0),\n        writexl (>= 1.3.1), xfun, xopen (>= 1.0.0), yesno (>= 0.1.2)\nRemotes: github::r-lib/rprojroot#79, gitlab::salim_b/r/pkgs/pkgsnip\nConfig/testthat/edition: 3\nEncoding: UTF-8\nLazyData: true\nRoxygen: list(markdown = TRUE)\nRoxygenNote: 7.1.1\nAuthor: Salim Brüggemann [aut, cre] (<https://orcid.org/0000-0002-5329-5987>)\nMaintainer: Salim Brüggemann <salim-b@pm.me>\nBuilt: R 4.0.4; ; 2021-02-28 14:41:28 UTC; unix\n