Read in a text file from a GitHub repository

Downloads the text file under the specified path from a GitHub repository via GitHub's GraphQL API v4 and returns its content as a string.

gh_text_file(path, owner, name, branch = "main")

Arguments

path
The path from the repository's root to the desired directory or file. A path or character scalar.

owner
The repository owner's GitHub user or organization name. A character scalar.

name
The repository name. A character scalar.

branch
The repository's desired branch name. A character scalar.

Value

A character scalar, or an empty character vector in case no text file is found under branch:path.

Details

Works for both public and private repositories, for the latter you just need to set up a sufficiently authorized GitHub Personal Access Token (PAT).

Note that nothing is returned in case of a binary file, as if no file at all existed under the given path.

See also

Examples

pal::gh_text_file(path = "pal.Rproj",
                  owner = "salim-b",
                  name = "pal",
                  branch = "master") |>
  cat()

#> Version: 1.0
#> 
#> RestoreWorkspace: No
#> SaveWorkspace: No
#> AlwaysSaveHistory: Default
#> 
#> EnableCodeIndexing: Yes
#> UseSpacesForTab: Yes
#> NumSpacesForTab: 2
#> Encoding: UTF-8
#> 
#> RnwWeave: Sweave
#> LaTeX: pdfLaTeX
#> 
#> AutoAppendNewline: Yes
#> StripTrailingWhitespace: Yes
#> LineEndingConversion: Posix
#> 
#> BuildType: Package
#> PackageUseDevtools: Yes
#> PackageInstallArgs: --no-multiarch --with-keep.source
#> PackageRoxygenize: rd,collate,namespace
#> 
#> QuitChildProcessesOnExit: Yes
#> 
#> MarkdownWrap: Column
#> MarkdownWrapAtColumn: 160
#> MarkdownReferences: Section
#> MarkdownCanonical: Yes