Read in text files from a GitHub repository

Downloads all text files under the specified path from a GitHub repository via GitHub's GraphQL API v4 and returns a named character vector with the file paths as names and the file contents as values.

This is a simple convenience function combining gh_dir_ls() and gh_text_file().

gh_text_files(path, owner, name, rev = "HEAD", recurse = FALSE)

Arguments

path

Path from the repository's root to the desired directory. A path or character scalar.

owner

Repository owner's GitHub user or organization name. A character scalar.

name

Repository name. A character scalar.

rev

Git revision expression matching the desired Git tree object, e.g. a branch name or another symbolic reference like "HEAD@{yesterday}" or "HEAD~10". A character scalar.

recurse

Whether or not to also include text files in subfolders of path. A logical scalar. Enabling this option may result in many API calls and thus produce a significant delay.

Value

A named character vector of length equal to the number of files found under rev:path with the file paths as names and the file contents as values.

Details

Works for both public and private repositories, for the latter you just need to set up a sufficiently authorized GitHub Personal Access Token (PAT).

Note that nothing is returned in case of a binary file, as if no file at all existed under the given path.

See also

Examples

pal::gh_text_files(path = "tests",
                   owner = "salim-b",
                   name = "pal") |>
  str()
#>  Named chr "library(testthat)\nlibrary(pal)\n\ntest_check(\"pal\")\n"
#>  - attr(*, "names")= chr "tests/testthat.R"

# you have to opt-in into directory recursion
pal::gh_text_files(path = "tests",
                   owner = "salim-b",
                   name = "pal",
                   recurse = TRUE) |>
  str()
#>  Named chr [1:2] "library(testthat)\nlibrary(pal)\n\ntest_check(\"pal\")\n" ...
#>  - attr(*, "names")= chr [1:2] "tests/testthat.R" "tests/testthat/test-pal.R"