Test if a directory is an R package

Convenience wrapper around the rprojroot::is_r_package root criterion. Note that it will only return TRUE for the root of a package directory, not its subdirectories.

is_pkg_dir(path = ".")

Arguments

path
The path of the directory to check. A character scalar. Defaults to the current working directory.

Value

TRUE if path is the root directory of an R package, FALSE otherwise.

See also

Examples

is_pkg_dir()

#> [1] FALSE

is_pkg_dir(fs::path_package("pal"))

#> [1] TRUE