Parse (R) Markdown as CommonMark XML tree

Parses (R) Markdown file content and returns it as a CommonMark parse tree in XML format.

md_xml(
  md,
  smart_punctuation = TRUE,
  hardbreaks = FALSE,
  normalize = TRUE,
  sourcepos = FALSE,
  extensions = TRUE,
  eol = c("LF", "CRLF", "CR", "LFCR")
)

Arguments

md
The (R) Markdown file content as a character scalar.

smart_punctuation
Whether to enable Pandoc's smart extension which converts straight quotes to curly quotes, --- to an em-dash (—), -- to an en-dash (–), and ... to ellipses (…). Nonbreaking spaces are inserted after certain abbreviations, such Mr..

hardbreaks
Whether or not to treat newlines as hard line breaks.

normalize
Consolidate adjacent text nodes.

sourcepos
Include source position attribute in output.

extensions
Enables Github extensions. Can be TRUE (all) FALSE (none) or a character vector with a subset of available extensions.

eol
End of line (EOL) control character sequence. One of

"LF" for the line feed (LF) character ("\n"). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the default.

"CRLF" for the carriage return + line feed (CR+LF) character sequence ("\r\n"). The standard on Microsoft Windows, DOS and some other systems.

"CR" for the carriage return (CR) character ("\r"). The standard on classic Mac OS and some other antiquated systems.

"LFCR" for the line feed + carriage return (LF+CR) character sequence ("\n\r"). The standard on RISC OS and some other exotic systems.

Value

An xml_node or xml_nodeset (possibly empty). Results are always de-duplicated.

Examples

pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
  pal::md_xml()

#> {xml_nodeset (205)}
#>  [1] <heading level="1">\n  <text xml:space="preserve">INTERNAL</text>\n</hea ...
#>  [2] <heading level="2">\n  <text xml:space="preserve">Logo</text>\n</heading>
#>  [3] <paragraph>\n  <text xml:space="preserve">The package logo is based on t ...
#>  [4] <paragraph>\n  <text xml:space="preserve">This was fed to </text>\n  <li ...
#>  [5] <code_block info="{r, purl = FALSE}" xml:space="preserve">hexSticker::st ...
#>  [6] <paragraph>\n  <text xml:space="preserve">Finally, </text>\n  <code xml: ...
#>  [7] <heading level="2">\n  <text xml:space="preserve">Avoid </text>\n  <code ...
#>  [8] <paragraph>\n  <text xml:space="preserve">Cf. </text>\n  <link destinati ...
#>  [9] <code_block info="{r}" xml:space="preserve">utils::globalVariables(names ...
#> [10] <heading level="2">\n  <text xml:space="preserve">Constants</text>\n</he ...
#> [11] <code_block info="{r}" xml:space="preserve"># forbidden dots arguments\n ...
#> [12] <heading level="2">\n  <text xml:space="preserve">Functions</text>\n</he ...
#> [13] <heading level="3">\n  <code xml:space="preserve">normalize_tree_path</c ...
#> [14] <paragraph>\n  <text xml:space="preserve">Normalize a tree path. Useful  ...
#> [15] <code_block info="{r}" xml:space="preserve">normalize_tree_path &lt;- fu ...
#> [16] <heading level="1">\n  <text xml:space="preserve">Statistical computing  ...
#> [17] <heading level="2">\n  <code xml:space="preserve">safe_seq_len</code>\n< ...
#> [18] <code_block info="{r}" xml:space="preserve">#' Generate an integer seque ...
#> [19] <heading level="2">\n  <code xml:space="preserve">safe_max</code>\n</hea ...
#> [20] <paragraph>\n  <text xml:space="preserve">Note that the </text>\n  <code ...
#> ...