Create a pal

Pals are persistent LLM-driven helpers designed to help you complete common tasks in interactive data analysis, authoring, and package development. Once created, they can be attached to a keybinding and immediately get to work on repetitive but hard-to-automate tasks.

To create a pal, simply pass pal() the ID of a pre-defined pal and a keybinding you'd like it attached to. For example, to use the cli pal:

pal("cli", "Ctrl+Shift+C")

Usage

pal(
  role = NULL,
  keybinding = NULL,
  fn = getOption(".pal_fn", default = "new_chat_claude"),
  ...,
  .ns = "elmer"
)

Arguments

role

The identifier for a pal prompt. Currently one of "cli" or "testthat".

keybinding

A key binding for the pal.

fn

A new_*() function, likely from the elmer package. Defaults to elmer::new_chat_claude(). To set a persistent alternative default, set the .pal_fn option; see examples below.

...

Additional arguments to fn. The system_prompt argument will be ignored if supplied. To set persistent defaults, set the .pal_args option; see examples below.

.ns

The package that the new_*() function is exported from.

Details

Upon successfully creating a pal, this function will assign the result to the search path as .last_pal. At that point, .pal_cli() and the RStudio add-in "Convert to cli" know to look for .last_pal and you don't need to worry about passing your cli pal yourself.

If you have an Anthropic API key (or another API key and the pal_*() options) set and this package installed, you are ready to using the add-in in any R session with no setup or library loading required; the addin knows to look for your API credentials and will call both this function and .pal_cli() itself.

Examples

if (FALSE) {
# to create a chat with claude:
pal()

# or with OpenAI's 4o-mini:
pal(
  "new_chat_openai",
  model = "gpt-4o-mini"
)

# to set OpenAI's 4o-mini as the default, for example, set the
# following options (possibly in your .Rprofile, if you'd like
# them to persist across sessions):
options(
  .pal_fn = "new_chat_openai",
  .pal_args = list(model = "gpt-4o-mini")
)
}