The pal directory

Pals can be added in batch with .pal_add_dir(), which takes a directory of markdown files. On package load, pal will add all pals defined in the pal directory, located at .pal_dir() (by default, ~/.config/pal). To add a new pal that will be loaded every time the package is loaded, use .pal_new().

Usage

.pal_add_dir(dir)

.pal_dir()

.pal_new(role, interface)

Arguments

dir

Path to a directory of markdown files–see Details for more.

role

A single string giving a descriptor of the pal's functionality.

interface

One of "replace", "prefix", or "suffix", describing how the pal will interact with the selection. For example, the cli pal "replace"s the selection, while the roxygen pal "prefixes" the selected code with documentation.

Details

Prompts are markdown files with the name role-interface.md, where interface is one of "replace", "prefix" or "suffix". An example directory might look like:

/
├── .config/
│   └── pal/
│       ├── proofread-replace.md
│       └── summarize-prefix.md

In that case, pal will register two custom pals when you call library(pal). One of them has the role "proofread" and will replace the selected text with a proofread version (according to the instructions contained in the markdown file itself). The other has the role "summarize" and will prefix the selected text with a summarized version (again, according to the markdown file's instructions). Note:

Files without a .md extension are ignored.

Files with a .md extension must contain only one hyphen in their filename, and the text following the hyphen must be one of replace, prefix, or suffix.

To load custom prompts every time the package is loaded, place your prompts in ~/.config/pal (or, to use some other folder, set options(.pal_dir = some_dir) before loading the package).