Creating custom pals

Users can create custom pals using the .pal_add() function; after passing the function a role and prompt, the pal will be available on the command palette.

Usage

.pal_add(role, prompt = NULL, interface = c("replace", "prefix", "suffix"))

.pal_remove(role)

Arguments

role

A single string giving a descriptor of the pal's functionality.

prompt

A file path to a markdown file giving the system prompt or the output of elmer::interpolate().

interface

One of "replace", "prefix", or "suffix", describing how the pal will interact with the selection. For example, the cli pal "replace"s the selection, while the roxygen pal "prefixes" the selected code with documentation.

dir

A directory of markdown files. See "Adding multiple, persistent pals" section below.

Value

NULL, invisibly. Called for its side effect: a pal with role role is registered with the pal package.

Adding multiple, persistent pals