Get package configuration value

Retrieves a package configuration value in a canonical way. The following configuration sources are consulted in descending order and the first hit is returned:

The R option <pkg>.<key>.

The environment variable <PKG>_<KEY>.

The configuration's default value as specified in the package's configuration metadata (<pkg>::pkg_config$default_value). If no default value is specified (NULL), an error is thrown.

pkg_config_val(key, pkg, default = NULL)

Arguments

key

Configuration key name. A character scalar.

pkg

Package name. A character scalar.

default

Default value to fall back to if neither the R option <pkg>.<key> nor the environment variable <PKG>_<KEY> is set. If NULL, the default value for key in <pkg>::pkg_config will be used (if defined).

Value

pkgsnip::return_label("r_obj")

Details

This function is intended to be used by package authors who want to expose their package configuration options in a canonical way (as outlined above). For pkg_config_val() to properly work, the configuration metadata must be available in the package's namespace as object pkg_config, which must be a dataframe or tibble with at minimum the columns key (of type character holding the configuration key names) and default_value (of type list holding the default configuration values, if any).

See also