Print package configuration metadata

Prints a package's configuration metadata (<pkg>::pkg_config) as a prettily formatted Markdown table.

Usage

print_pkg_config(pkg, roxy_to_md = FALSE)

Arguments

pkg

Package name. A character scalar.

roxy_to_md

Whether or not to convert roxygen2 documentation links in pseudo-Markdown style to actual Markdown ones using roxy_to_md_links().

Value

A character vector.

See also

Examples

try(
  pal::print_pkg_config(pkg = "pkgpurl")
)
#> 
#> 
#> |**Description**                                                                                                                                                                                                                                                                                                                                                                                                                                                 |**R option**                   |**Environment variable**         |**Default value** |
#> |:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:------------------------------|:--------------------------------|:-----------------|
#> |Whether or not to add a **copyright notice** at the beginning of the generated `.R` files as recommended by e.g. the [GNU licenses](https://www.gnu.org/licenses/gpl-howto.html). The notice consists of the name and description of the program and the word `Copyright (C)` followed by the release years and the name(s) of the copyright holder(s), or if not specified, the author(s). The year is always the current year. All the other information is extracted from the package's `DESCRIPTION` file.|`pkgpurl.add_copyright_notice` |`R_PKGPURL_ADD_COPYRIGHT_NOTICE` |                  |
#> |Whether or not to add a **license notice** at the beginning of the generated `.R` files as recommended by e.g. the [GNU licenses](https://www.gnu.org/licenses/gpl-howto.html). The license is determined from the package's `DESCRIPTION` file and currently only the [`AGPL-3.0-or-later` license](https://spdx.org/licenses/AGPL-3.0-or-later.html) is supported.                                                                                            |`pkgpurl.add_license_notice`   |`R_PKGPURL_ADD_LICENSE_NOTICE`   |                  |
#> |Whether or not to overwrite [pkgdown](https://pkgdown.r-lib.org/)'s [reference index](https://pkgdown.r-lib.org/reference/build_reference.html#reference-index) in the configuration file `_pkgdown.yml` with an auto-generated one based on the main input file as described in [pkgpurl::gen_pkgdown_ref()].                                                                                                                                                  |`pkgpurl.gen_pkgdown_ref`      |`R_PKGPURL_GEN_PKGDOWN_REF`      |                  |