List a function's default parameter values in prose-style

Extracts the default value(s) of a function's definition and returns it in prose style listing.

prose_ls_fn_param(
  param,
  fn = sys.function(sys.parent()),
  env = parent.frame(),
  as_scalar = TRUE,
  wrap = "`",
  separator = ",",
  last_separator = " or "
)

Arguments

param
The parameter name. A character scalar.

fn
A function or a function name (searched for in env). See formals() for details.

env
The environment fn is defined in. See formals() for details.

as_scalar
Whether to return the result as a single string concatenated by separator and last_separator.

wrap
The string (usually a single character) in which params default values are to be wrapped.

separator
The separator to delimit params default values. Only relevant if as_scalar = TRUE.

last_separator
The separator to delimit the second-last and last one of params default values. Only relevant if as_scalar = TRUE.

Value

A character vector. Of length 1 if as_scalar = TRUE.

Details

This function can be very convenient to avoid duplication in roxygen2 documentation by leveraging inline R code evaluation as follows:

#' @param some_param Some parameter. One of `r pal::prose_ls_fn_param(param = "some_param", fn = "some_fn")`.
some_fn <- function(some_param = c("a", "b", "c")) {
  some_param <- rlang::arg_match(some_param)
  ...
}

Or to list the possible parameter values formatted as an unnumbered list instead, use the inline code `r pal::prose_ls_fn_param(param = "some_param", fn = "some_fn", as_scalar = FALSE) %>% pal::as_md_list()` in the example above.

Caveats

This function does not work for Primitives.

deparse() is used internally to get a character representation of non-character default values. Therefore all of deparse()'s fuzziness also applies to this function.

See also

Examples

prose_ls_fn_param(param = ".name_repair",
                  fn = tibble::as_tibble) %>%
cat_lines()

#> `"check_unique"`,`"unique"`,`"universal"` or `"minimal"`

prose_ls_fn_param(param = ".name_repair",
                  fn = tibble::as_tibble,
                  as_scalar = FALSE) %>%
cat_lines()

#> `"check_unique"`
#> `"unique"`
#> `"universal"`
#> `"minimal"`