Get an object's roxygen2 tag value

Parses text for roxygen2 blocks and extracts the value(s) belonging to the tag_names documenting obj_name.

roxy_tag_value(text, obj_name, tag_name = "param", param_name = NULL)

Arguments

text
The R source code to extract the object's roxygen2 tag value from. A character vector.

obj_name
The object name to which the roxygen2 tag belongs to, usually a function name. A character scalar.

tag_name
The name of the roxygen2 tag (without the @) to extract the value from. A character scalar.

param_name
The parameter name to extract the value from. Only relevant if tag_name = "param". A character scalar.

Value

A character scalar if tag_name = "param" and param_name != NULL, otherwise a list.

Examples

text <- readr::read_lines(paste0("https://raw.githubusercontent.com/r-lib/rlang/",
                                 "db52a58d505b65f58ba922d4752b5b0061f2a98c/R/fn.R"))

roxy_tag_value(text = text,
               obj_name = "as_function",
               param_name = "x")

#> [1] "A function or formula.\n\n  If a **function**, it is used as is.\n\n  If a **formula**, e.g. `~ .x + 2`, it is converted to a function\n  with up to two arguments: `.x` (single argument) or `.x` and `.y`\n  (two arguments). The `.` placeholder can be used instead of `.x`.\n  This allows you to create very compact anonymous functions (lambdas) with up\n  to two inputs. Functions created from formulas have a special\n  class. Use `is_lambda()` to test for it.\n\n  Lambdas currently do not support [nse-force],\n  due to the way the arguments are handled internally."