Minimum (safe)

Modified version of min() that differs in the following ways:

NAs in the input (...) are ignored by default (rm_na = TRUE).

If the input is of length zero, the output will also be of length zero (of the same type as the input).

It is ensured that all inputs are either numeric, of length zero or NA. There is no case where the return value will be Inf.

Dynamic dots are supported.

safe_min(..., rm_na = TRUE)

Arguments

...
The numeric objects of which to determine the minimum. Dynamic dots are supported.

rm_na
Ignore missing values in .... If missing values are present and rm_na = FALSE, the result will always be NA.

Value

A numeric scalar or empty value, of the same type as ....

See also

Examples

# other than `base::min()`, this function removes `NA`s by default
min(1, NA, 2)

#> [1] NA

safe_min(1, NA, 2)

#> [1] 1

# other than `base::min()`, this function does not return `Inf` or `NA_character_` for
# zero-length inputs
if (FALSE) {
min(NULL)
min(character())
min(integer())
safe_min(NULL)
safe_min(character())
safe_min(integer())

# other than `base::min()`, this function fails for non-numeric, non-zero-length inputs
min("zero", 1L)
min("zero", "one")
safe_min("zero", 1L)
safe_min("zero", "one")}