Strip Markdown formatting from character vector

Removes all Markdown formatting from a character vector.

Usage

strip_md(x, strip_footnotes = TRUE)

Arguments

x

A character vector to strip Markdown formatting from.

strip_footnotes

Whether to remove Markdown footnotes, too. If FALSE, footnotes are canonicalized (to have sequential integer identifiers).

Value

A character vector of the same length as x.

Details

This function relies on commonmark::markdown_text() which supports the CommonMark specification plus the Github extensions. Unfortunately, Markdown footnotes can't be stripped using commonmark::markdown_text(). Therefore a separate option strip_footnotes is offered which relies on regular expressions to remove inline footnotes and footnote definitions and references.

See also

Examples

pal::strip_md(
  "A **MD** formatted [string](https://en.wikipedia.org/wiki/String_(computer_science))"
)
#> [1] "A MD formatted string"

# link references are only removed *iff* the reference is included in `x`:
pal::strip_md("[A reference link][refid]\n\n[refid]: https://example.com")
#> [1] "A reference link"
pal::strip_md("[A reference link][refid]\n\n_No ref here..._")
#> [1] "[A reference link][refid]\n\nNo ref here..."