Assert MIME type

Asserts a response object is of a specific MIME type. Convenience wrapper around httr::http_type().

assert_mime_type(
  response,
  mime_type,
  msg = paste0("The response's MIME type is {.val {mime_type_actual}} ",
    "but expected was {.val {mime_type}}."),
  msg_suffix = ""
)

Arguments

response
A response object.

mime_type
Expected MIME type, e.g. "application/json". A character scalar.

msg
Message to display in case of an error. Supports cli's inline markup. A character scalar.

msg_suffix
Additional string to append to msg. Supports cli's inline markup.

Value

response, invisibly.

See also

Examples

httr::GET("https://api.github.com/users/salim-b") |>
  pal::assert_mime_type("application/json") |>
  httr::content()
#> $login
#> [1] "salim-b"
#> 
#> $id
#> [1] 20040931
#> 
#> $node_id
#> [1] "MDQ6VXNlcjIwMDQwOTMx"
#> 
#> $avatar_url
#> [1] "https://avatars.githubusercontent.com/u/20040931?v=4"
#> 
#> $gravatar_id
#> [1] ""
#> 
#> $url
#> [1] "https://api.github.com/users/salim-b"
#> 
#> $html_url
#> [1] "https://github.com/salim-b"
#> 
#> $followers_url
#> [1] "https://api.github.com/users/salim-b/followers"
#> 
#> $following_url
#> [1] "https://api.github.com/users/salim-b/following{/other_user}"
#> 
#> $gists_url
#> [1] "https://api.github.com/users/salim-b/gists{/gist_id}"
#> 
#> $starred_url
#> [1] "https://api.github.com/users/salim-b/starred{/owner}{/repo}"
#> 
#> $subscriptions_url
#> [1] "https://api.github.com/users/salim-b/subscriptions"
#> 
#> $organizations_url
#> [1] "https://api.github.com/users/salim-b/orgs"
#> 
#> $repos_url
#> [1] "https://api.github.com/users/salim-b/repos"
#> 
#> $events_url
#> [1] "https://api.github.com/users/salim-b/events{/privacy}"
#> 
#> $received_events_url
#> [1] "https://api.github.com/users/salim-b/received_events"
#> 
#> $type
#> [1] "User"
#> 
#> $site_admin
#> [1] FALSE
#> 
#> $name
#> [1] "Salim B"
#> 
#> $company
#> NULL
#> 
#> $blog
#> [1] "https://keybase.io/salimb"
#> 
#> $location
#> [1] "Switzerland"
#> 
#> $email
#> NULL
#> 
#> $hireable
#> NULL
#> 
#> $bio
#> NULL
#> 
#> $twitter_username
#> NULL
#> 
#> $public_repos
#> [1] 60
#> 
#> $public_gists
#> [1] 3
#> 
#> $followers
#> [1] 13
#> 
#> $following
#> [1] 1
#> 
#> $created_at
#> [1] "2016-06-20T09:11:58Z"
#> 
#> $updated_at
#> [1] "2022-07-29T14:08:01Z"
#> 

# an informative error is thrown when the assertion is violated
try(
  httr::GET("https://api.github.com/users/salim-b") |> pal::assert_mime_type("text/plain")
)
#> Error in pal::assert_mime_type(httr::GET("https://api.github.com/users/salim-b"),  : 
#>   The response's MIME type is "application/json" but expected was
#> "text/plain".