Assert a package is installed

Assert a package is installed

assert_pkg(pkg, min_version = NULL, message = NULL, install_hint = NULL)

Arguments

pkg
Package name. A character scalar.

min_version
Minimum required version number of pkg. Must be either NULL to ignore version numbers, or a single package_version or something coercible to.

message
The error message to display in case the package is not installed. If NULL, a sensible standard message is generated.

install_hint
Additional package installation instructions appended to message. Either NULL in order to autogenerate the hint, or a character scalar. Set install_hint = "" in order to disable the hint.

Value

pkg invisibly.

See also

Examples

assert_pkg("pal")

assert_pkg(pkg = "glue",
           message = paste0("You should really consider to install the awesome `glue` package! ",
                            "It's the glue that keeps strings and variables together 🤲."))
if (FALSE) {
assert_pkg("pal", min_version = 9999.9)

assert_pkg("yay",
           install_hint = paste0("To install the latest development version, run ",
                                 "`remotes::install_gitlab(\"salim_b/r/pkgs/yay\")`."))}