Build README.Rmd

A simpler, but considerably faster alternative to devtools::build_readme() since it doesn't install your package in a temporary library before building the README.Rmd. This has the pleasant side effect that, other than the latter function, it also works for .Rmd files which aren't part of an R package.

build_readme(
  input = "README.Rmd",
  output = "README.md",
  build_index_md = NULL,
  env = parent.frame()
)

Arguments

input
The path to the R Markdown README file to be built. A character scalar.

output
The path of the built Markdown README. A character scalar.

build_index_md
Whether to build a separate pkgdown-optimized pkgdown/index.md alongside output (i.e. in the same parent directory). If NULL, it will only be built if the parent directory of output contains a pkgdown configuration file. Note that it will be built with the R option pal.build_readme.is_pkgdown = TRUE, allowing for conditional content inclusion in input – e.g. via the code chunk option eval = isTRUE(getOption("pal.build_readme.is_pkgdown")).

env
Environment in which code chunks are to be evaluated, e.g. parent.frame(), new.env(), or globalenv().

Value

The path to input as a character scalar, invisibly.

Details

Note that for public package repositories, it's recommended to use devtools::build_readme() since it ensures the README.Rmd can be built reproducibly, which means all the objects and files it references must be accessible from the repository.

This function is also registered as an RStudio add-in allowing RStudio users to assign a custom shortcut to it.

See also