Check that all dot parameter names are a valid subset of a function's parameter names.

Ensures that dots (...) are either empty (if .empty_ok = TRUE), or all named dot parameter names are a valid subset of a function's parameter names. In case of an invalid or .forbidden argument, an informative message is shown and the defined .action is taken.

check_dots_named(
  ...,
  .function,
  .additional = NULL,
  .forbidden = NULL,
  .empty_ok = TRUE,
  .action = rlang::abort
)

Arguments

...
The dots argument to check.

.function
The function the ... will be passed on to.

.additional
Parameter names within ... that should be treated as valid in addition to .function's actual parameter names. A character vector.

.forbidden
Parameter names within ... that should be treated as invalid. This has precedence over .additional. A character vector.

.empty_ok
Set to TRUE if empty ... should be allowed, or to FALSE otherwise.

.action
The action to take when the check fails. One of

rlang::abort

rlang::warn

rlang::inform

rlang::signal

Details

check_dots_named() is intended to combat the second one of the two major downsides that using ... usually brings. In chapter 6.6 of the book Advanced R it is phrased as follows:

Using ... comes with two downsides:

When you use it to pass arguments to another function, you have to carefully explain to the user where those arguments go. This makes it hard to understand what you can do with functions like lapply() and plot().

A misspelled argument will not raise an error. This makes it easy for typos to go unnoticed.

Examples

# We can use `check_dots_named()` to address this second downside:
sum_safe <- function(...,
                     na.rm = FALSE) {
  pal::check_dots_named(...,
                        .function = sum)
  sum(...,
      na.rm = na.rm)
}

# note how the misspelled `na_rm` (instead of `na.rm`) silently gets ignored
# in the original function
sum(1, 2, NA, na_rm = TRUE)

#> [1] NA

if (FALSE) {
# whereas our safe version properly errors
sum_safe(1, 2, NA, na_rm = TRUE)}

# we can even build an `sapply()` function that fails "intelligently" 
sapply_safe <- function(X,
                        FUN,
                        ...,
                        simplify = TRUE,
                        USE.NAMES = TRUE) {
  pal::check_dots_named(...,
                        .function = FUN)
  sapply(X = X,
         FUN = FUN,
         ...,
         simplify = TRUE,
         USE.NAMES = TRUE)
}

# while the original `sapply()` silently ignores misspelled arguments,
sapply(1:5, paste, "hour workdays", sep = "-", colaspe = " ")

#> [1] "1-hour workdays- " "2-hour workdays- " "3-hour workdays- "
#> [4] "4-hour workdays- " "5-hour workdays- "

if (FALSE) {
# `sapply_safe()` will throw an informative error message
sapply_safe(1:5, paste, "hour workdays", sep = "-", colaspe = " ")}

if (FALSE) {
# but be aware that `check_dots_named()` might be a bit rash
sum_safe(a = 1, b = 2)}

# while the original function actually has nothing to complain
sum(a = 1, b = 2)

#> [1] 3

if (FALSE) {
# also, it doesn't play nicely with functions that don't expose all of
# their arg names (`to` and `by` in the case of `seq()`)
sapply_safe(X = c(0,50),
            FUN = seq,
            to = 100,
            by = 5)}

# but providing `to` and `by` *unnamed* is fine of course:
sapply_safe(X = c(0,50),
            FUN = seq,
            100,
            5)

#> [[1]]
#>  [1]   0   5  10  15  20  25  30  35  40  45  50  55  60  65  70  75  80  85  90
#> [20]  95 100
#> 
#> [[2]]
#>  [1]  50  55  60  65  70  75  80  85  90  95 100
#>