Create readr column specification using regular expression matching

Allows to define a regular expression per desired column specification object matching the respective column names.

cols_regex(..., .default = readr::col_character(), .col_names)

Arguments

...
Named arguments where the names are (Perl-compatible) regular expressions and the values are column objects created by col_*(), or their abbreviated character names (as described in the col_types parameter of readr::read_delim())Dynamic dots are supported.)`

.default
Any named columns not matched by any of the regular expressions in ... will be read with this column type.

.col_names
Column names which should be matched by ....

Value

A column specification.

Examples

library(magrittr)

# for some hypothetical CSV data column names like these ...
col_names <- c("VAR1_Text",
               "VAR2_Text",
               "VAR3_Text_Other",
               "VAR1_Code_R1",
               "VAR2_Code_R2",
               "HAS_R1_Lag",
               "HAS_R2_Lag",
               "GARBAGEX67",
               "GARBAGEY09")

# ... a column spec could be created concisely as follows:
col_regex <- list("_Text(_|$)" = "c",
                  "_Code(_|$)" = "i",
                  "^GARBAGE"  = readr::col_skip())

pal::cols_regex(.col_names = col_names,
                !!!col_regex,
                .default     = "l")
#> cols(
#>   .default = col_logical(),
#>   VAR1_Text = col_character(),
#>   VAR2_Text = col_character(),
#>   VAR3_Text_Other = col_character(),
#>   VAR1_Code_R1 = col_integer(),
#>   VAR2_Code_R2 = col_integer(),
#>   GARBAGEX67 = col_skip(),
#>   GARBAGEY09 = col_skip()
#> )

# we can parse some real data:
raw_data <-
  httr::GET(paste0("http://www.web.statistik.zh.ch/ogd/data/",
                   "KANTON_ZUERICH_nrw_2019_listen_ergebnisse_gemeinde.csv")) %>%
  httr::content(as = "text",
                encoding = "UTF-8")

readr::read_csv(file = raw_data,
                col_types = pal::cols_regex("^(Gemeindenamen|Partei)$" = "c",
                                            "(?i)anteil" = "d",
                                            .default = "i",
                                            .col_names = pal::dsv_colnames(raw_data)))
#> # A tibble: 3,043 × 8
#>    Gemeindenamen `BFS-Nr.` Partei  Stimmen Wähler Wähleranteil `+/- (Wählerante…
#>    <chr>             <int> <chr>     <int>  <int>        <dbl>             <dbl>
#>  1 Adlikon              21 SVP        5241    150        56.5              -0.36
#>  2 Adlikon              21 SP          566     16         6.1              -2.22
#>  3 Adlikon              21 FDP         869     25         9.36              1.97
#>  4 Adlikon              21 glp         595     17         6.41              2.49
#>  5 Adlikon              21 CVP         123      4         1.33              0.46
#>  6 Adlikon              21 GRÜNE       628     18         6.77              2.93
#>  7 Adlikon              21 BDP         363     10         3.91             -2.72
#>  8 Adlikon              21 EVP         623     18         6.71              1.99
#>  9 Adlikon              21 Piraten       0      0         0                -0.43
#> 10 Adlikon              21 IP            1      0         0.01             NA   
#> # … with 3,033 more rows, and 1 more variable: Wähleranteil 2015 <dbl>

# an alternative way to process the same data using `readr::type_convert()`:
readr::read_csv(file = raw_data,
                col_types = list(.default = "c")) %>%
  readr::type_convert(col_types = pal::cols_regex("^(Gemeindenamen|Partei)$" = "c",
                                                  "(?i)anteil" = "d",
                                                  .default = "i",
                                                  .col_names = colnames(.)))
#> # A tibble: 3,043 × 8
#>    Gemeindenamen `BFS-Nr.` Partei  Stimmen Wähler Wähleranteil `+/- (Wählerante…
#>    <chr>             <int> <chr>     <int>  <int>        <dbl>             <dbl>
#>  1 Adlikon              21 SVP        5241    150        56.5              -0.36
#>  2 Adlikon              21 SP          566     16         6.1              -2.22
#>  3 Adlikon              21 FDP         869     25         9.36              1.97
#>  4 Adlikon              21 glp         595     17         6.41              2.49
#>  5 Adlikon              21 CVP         123      4         1.33              0.46
#>  6 Adlikon              21 GRÜNE       628     18         6.77              2.93
#>  7 Adlikon              21 BDP         363     10         3.91             -2.72
#>  8 Adlikon              21 EVP         623     18         6.71              1.99
#>  9 Adlikon              21 Piraten       0      0         0                -0.43
#> 10 Adlikon              21 IP            1      0         0.01             NA   
#> # … with 3,033 more rows, and 1 more variable: Wähleranteil 2015 <dbl>