Deploy pkgdown site to local Git folder

Copies the static pkgdown website files to another local Git folder, then stages, commits and pushes the changes. Use pkgdown::build_site() before running this function in order to create the website files.

Use this function with caution since it completely wipes the to_path directory by default!

deploy_pkgdown_site(
  pkg_path = ".",
  to_path,
  use_dev_build = NULL,
  clean_to_path = TRUE,
  commit_msg = "auto-deploy pkgdown"
)

Arguments

pkg_path
Path to the R package of which the pkgdown website files are to be deployed.

to_path
Path to the Git (sub)folder to which the pkgdown website files are to be deployed.

use_dev_build
Whether to deploy the development build of the pkgdown website files.

clean_to_path
Whether to completely wipe to_path before deploying the new pkgdown website files. Setting this TRUE ensures there are no obsolete files left over from previous deployments.

commit_msg
The Git commit message used for the deployment.

Value

A vector of paths to the deployed files/folders, invisibly.