Get all DESCRIPTION file fields as cleaned up list

Returns all fields from a specific DESCRIPTION file as a named list with values cleaned up:

Whitespaces at the start and end of field values as well as repeated whitespaces within them are removed.

Multi-value fields are returned as vectors.

The fields Depends, Imports and Suggests are returned as a single data frame named dependencies.

Usage

desc_list(file = ".")

Arguments

file

DESCRIPTION file to use. By default the DESCRIPTION file of the current package (i.e. the package the working directory is part of) is used.

Value

A list.

See also

Examples

fs::path_package(package = "dplyr") |> pal::desc_list()
#> $Type
#> [1] "Package"
#> 
#> $Package
#> [1] "dplyr"
#> 
#> $Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $Version
#> [1] "1.1.1"
#> 
#> $Description
#> [1] "A fast, consistent tool for working with data frame like objects, both in memory and out of memory."
#> 
#> $License
#> [1] "MIT + file LICENSE"
#> 
#> $BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $VignetteBuilder
#> [1] "knitr"
#> 
#> $`Config/Needs/website`
#> [1] "tidyverse, shiny, pkgdown, tidyverse/tidytemplate"
#> 
#> $`Config/testthat/edition`
#> [1] "3"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "true"
#> 
#> $RoxygenNote
#> [1] "7.2.3"
#> 
#> $NeedsCompilation
#> [1] "yes"
#> 
#> $Packaged
#> [1] "2023-03-21 21:01:28 UTC; hadleywickham"
#> 
#> $Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>), Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>), Lionel Henry [aut], Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>), Davis Vaughan [aut] (<https://orcid.org/0000-0003-4777-038X>), Posit Software, PBC [cph, fnd]"
#> 
#> $Maintainer
#> [1] "Hadley Wickham <hadley@posit.co>"
#> 
#> $Repository
#> [1] "CRAN"
#> 
#> $`Date/Publication`
#> [1] "2023-03-22 13:20:07 UTC"
#> 
#> $Built
#> [1] "R 4.2.3; x86_64-pc-linux-gnu; 2023-03-22 16:53:27 UTC; unix"
#> 
#> $`Authors@R`
#> [1] "Hadley Wickham <hadley@posit.co> [aut, cre] (<https://orcid.org/0000-0003-4757-117X>)"
#> [2] "Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>)"                      
#> [3] "Lionel Henry [aut]"                                                                   
#> [4] "Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>)"                        
#> [5] "Davis Vaughan <davis@posit.co> [aut] (<https://orcid.org/0000-0003-4777-038X>)"       
#> [6] "Posit Software, PBC [cph, fnd]"                                                       
#> 
#> $dependencies
#>        type        package  version
#> 1   Depends              R >= 3.5.0
#> 2   Imports            cli >= 3.4.0
#> 3   Imports       generics        *
#> 4   Imports           glue >= 1.3.2
#> 5   Imports      lifecycle >= 1.0.3
#> 6   Imports       magrittr   >= 1.5
#> 7   Imports        methods        *
#> 8   Imports         pillar >= 1.5.1
#> 9   Imports             R6        *
#> 10  Imports          rlang >= 1.1.0
#> 11  Imports         tibble >= 3.2.0
#> 12  Imports     tidyselect >= 1.2.0
#> 13  Imports          utils        *
#> 14  Imports          vctrs >= 0.6.0
#> 15 Suggests          bench        *
#> 16 Suggests          broom        *
#> 17 Suggests          callr        *
#> 18 Suggests           covr        *
#> 19 Suggests            DBI        *
#> 20 Suggests         dbplyr >= 2.2.1
#> 21 Suggests        ggplot2        *
#> 22 Suggests          knitr        *
#> 23 Suggests         Lahman        *
#> 24 Suggests         lobstr        *
#> 25 Suggests microbenchmark        *
#> 26 Suggests   nycflights13        *
#> 27 Suggests          purrr        *
#> 28 Suggests      rmarkdown        *
#> 29 Suggests         RMySQL        *
#> 30 Suggests    RPostgreSQL        *
#> 31 Suggests        RSQLite        *
#> 32 Suggests        stringi >= 1.7.6
#> 33 Suggests       testthat >= 3.1.5
#> 34 Suggests          tidyr >= 1.3.0
#> 35 Suggests          withr        *
#> 
#> $URL
#> [1] "https://dplyr.tidyverse.org"        "https://github.com/tidyverse/dplyr"
#>