Get all DESCRIPTION file fields as cleaned up list

Returns all fields from a specific DESCRIPTION file as a named list with values cleaned up:

Whitespaces at the start and end of field values as well as repeated whitespaces within them are removed.

Multi-value fields are returned as vectors.

The fields Depends, Imports and Suggests are returned as a single data frame named dependencies.

desc_list(file = ".")

Arguments

file
DESCRIPTION file to use. By default the DESCRIPTION file of the current package (i.e. the package the working directory is part of) is used.

Value

A list.

See also

Examples

try(
  pal::desc_list(file = fs::path_package(package = "pal"))
)
#> $Package
#> [1] "pal"
#> 
#> $Title
#> [1] "Friendly Convenience/Utility Functions"
#> 
#> $Version
#> [1] "0.0.9075"
#> 
#> $Description
#> [1] "A collection of convenience/utility functions complementing base R and the tidyverse."
#> 
#> $License
#> [1] "AGPL (>= 3)"
#> 
#> $BugReports
#> [1] "https://gitlab.com/salim_b/r/pkgs/pal/issues"
#> 
#> $Remotes
#> [1] "github::r-lib/rprojroot#79, github::ropensci/tinkr, gitlab::salim_b/r/pkgs/pkgsnip"
#> 
#> $`Config/testthat/edition`
#> [1] "3"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "true"
#> 
#> $Roxygen
#> [1] "list(markdown = TRUE)"
#> 
#> $RoxygenNote
#> [1] "7.1.2"
#> 
#> $Author
#> [1] "Salim Brüggemann [aut, cre] (<https://orcid.org/0000-0002-5329-5987>)"
#> 
#> $Maintainer
#> [1] "Salim Brüggemann <salim-b@pm.me>"
#> 
#> $Built
#> [1] "R 4.1.2; ; 2021-12-17 23:31:36 UTC; unix"
#> 
#> $`Authors@R`
#> [1] "Salim Brüggemann <salim-b@pm.me> [aut, cre] (<https://orcid.org/0000-0002-5329-5987>)"
#> 
#> $dependencies
#>        type     package       version
#> 1   Depends           R        >= 4.1
#> 2   Imports   checkmate      >= 2.0.0
#> 3   Imports         cli      >= 3.0.1
#> 4   Imports       dplyr      >= 1.0.7
#> 5   Imports          fs      >= 1.5.0
#> 6   Imports        glue      >= 1.4.2
#> 7   Imports   lifecycle      >= 1.0.0
#> 8   Imports    magrittr      >= 2.0.1
#> 9   Imports     methods             *
#> 10  Imports       purrr      >= 0.3.4
#> 11  Imports       rlang     >= 0.4.11
#> 12  Imports     stringr      >= 1.4.0
#> 13  Imports      tibble      >= 3.1.4
#> 14  Imports       utils             *
#> 15 Suggests        brio      >= 1.1.0
#> 16 Suggests  commonmark        >= 1.7
#> 17 Suggests        desc      >= 1.4.0
#> 18 Suggests    ellipsis      >= 0.3.2
#> 19 Suggests          gh      >= 1.3.0
#> 20 Suggests        httr      >= 1.4.1
#> 21 Suggests janeaustenr             *
#> 22 Suggests       knitr       >= 1.28
#> 23 Suggests      pander      >= 0.6.3
#> 24 Suggests     pkgdown >= 1.6.1.9000
#> 25 Suggests   pkgsearch      >= 3.0.3
#> 26 Suggests     pkgsnip   >= 0.0.9026
#> 27 Suggests    rappdirs      >= 0.3.3
#> 28 Suggests       readr      >= 1.4.0
#> 29 Suggests   rmarkdown        >= 2.1
#> 30 Suggests    roxygen2      >= 7.1.2
#> 31 Suggests   rprojroot      >= 1.3.2
#> 32 Suggests  rstudioapi       >= 0.11
#> 33 Suggests    testthat      >= 3.0.0
#> 34 Suggests       tinkr >= 0.0.0.9000
#> 35 Suggests       waldo      >= 0.2.3
#> 36 Suggests       withr      >= 2.4.1
#> 37 Suggests        xfun       >= 0.26
#> 38 Suggests        xml2      >= 1.3.2
#> 39 Suggests        xslt      >= 1.4.3
#> 
#> $URL
#> [1] "https://rpkg.dev/pal"                 
#> [2] "https://gitlab.com/salim_b/r/pkgs/pal"
#>