Get all DESCRIPTION file fields as cleaned up list

Returns all fields from a specific DESCRIPTION file as a named list with values cleaned up:

Whitespaces at the start and end of field values as well as repeated whitespaces within them are removed.

Multi-value fields are returned as vectors.

The fields Depends, Imports and Suggests are returned as a single data frame named dependencies.

desc_list(file = ".")

Arguments

file

DESCRIPTION file to use. By default the DESCRIPTION file of the current package (i.e. the package the working directory is part of) is used.

Value

A list.

See also

Examples

fs::path_package(package = "dplyr") |> pal::desc_list()
#> $Type
#> [1] "Package"
#> 
#> $Package
#> [1] "dplyr"
#> 
#> $Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $Version
#> [1] "1.0.10"
#> 
#> $Description
#> [1] "A fast, consistent tool for working with data frame like objects, both in memory and out of memory."
#> 
#> $License
#> [1] "MIT + file LICENSE"
#> 
#> $BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $VignetteBuilder
#> [1] "knitr"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "true"
#> 
#> $RoxygenNote
#> [1] "7.2.1"
#> 
#> $`Config/testthat/edition`
#> [1] "3"
#> 
#> $`Config/Needs/website`
#> [1] "tidyverse, shiny, pkgdown, tidyverse/tidytemplate"
#> 
#> $NeedsCompilation
#> [1] "yes"
#> 
#> $Packaged
#> [1] "2022-08-31 18:31:04 UTC; hadleywickham"
#> 
#> $Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>), Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>), Lionel Henry [aut], Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>), RStudio [cph, fnd]"
#> 
#> $Maintainer
#> [1] "Hadley Wickham <hadley@rstudio.com>"
#> 
#> $Repository
#> [1] "CRAN"
#> 
#> $`Date/Publication`
#> [1] "2022-09-01 09:20:06 UTC"
#> 
#> $Built
#> [1] "R 4.2.1; x86_64-pc-linux-gnu; \"Sun, 04 Sep 2022 00:42:58 +0000\"; unix"
#> 
#> $`Authors@R`
#> [1] "Hadley Wickham <hadley@rstudio.com> [aut, cre] (<https://orcid.org/0000-0003-4757-117X>)"
#> [2] "Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>)"                         
#> [3] "Lionel Henry [aut]"                                                                      
#> [4] "Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>)"                           
#> [5] "RStudio [cph, fnd]"                                                                      
#> 
#> $dependencies
#>        type        package  version
#> 1   Depends              R >= 3.4.0
#> 2   Imports       generics        *
#> 3   Imports           glue >= 1.3.2
#> 4   Imports      lifecycle >= 1.0.1
#> 5   Imports       magrittr   >= 1.5
#> 6   Imports        methods        *
#> 7   Imports             R6        *
#> 8   Imports          rlang >= 1.0.2
#> 9   Imports         tibble >= 2.1.3
#> 10  Imports     tidyselect >= 1.1.1
#> 11  Imports          utils        *
#> 12  Imports          vctrs >= 0.4.1
#> 13  Imports         pillar >= 1.5.1
#> 14 Suggests          bench        *
#> 15 Suggests          broom        *
#> 16 Suggests          callr        *
#> 17 Suggests           covr        *
#> 18 Suggests            DBI        *
#> 19 Suggests         dbplyr >= 1.4.3
#> 20 Suggests        ggplot2        *
#> 21 Suggests          knitr        *
#> 22 Suggests         Lahman        *
#> 23 Suggests         lobstr        *
#> 24 Suggests microbenchmark        *
#> 25 Suggests   nycflights13        *
#> 26 Suggests          purrr        *
#> 27 Suggests      rmarkdown        *
#> 28 Suggests         RMySQL        *
#> 29 Suggests    RPostgreSQL        *
#> 30 Suggests        RSQLite        *
#> 31 Suggests       testthat >= 3.1.1
#> 32 Suggests          tidyr        *
#> 33 Suggests          withr        *
#> 
#> $URL
#> [1] "https://dplyr.tidyverse.org"        "https://github.com/tidyverse/dplyr"
#>