Get the value from a DESCRIPTION file field, cleaned up and with dynamic fallback

Returns the value from a specific DESCRIPTION file field (aka key). Whitespaces at the start and end of the value as well as repeated whitespaces within it are removed.

Usage

desc_value(
  key,
  default = glue::glue("<No `{key}` field set in DESCRIPTION!>"),
  file = "."
)

Arguments

key

The field to query.

default

Default value to return if key is not found.

file

DESCRIPTION file to use. By default the DESCRIPTION file of the current package (i.e. the package the working directory is part of) is used.

Value

A character scalar.

Details

This function is a slightly modified version of desc::desc_get_field() that allows the default parameter to be dependent on the key parameter.

By default, the following string is returned if key = "NoRealKey" is not found:

"<No `NoRealKey` field set in DESCRIPTION!>"

If you rather want to take an action like throwing an error, it's recommended to call desc::desc_get_field() directly.

See also

Examples

pal::desc_value(key = "Description",
                file = fs::path_package("pal"))
#> [1] "A collection of convenience/utility functions complementing base R and the tidyverse."