Drop list elements by name

Drops all list elements whose names match the provided set.

Usage

list_drop(x, drop)

Arguments

x

A list.

drop

Names of the list elements to drop.

Value

A list.

See also

Examples

as.list(mtcars) |> pal::list_drop(c("disp", "drat", "qsec"))
#> $mpg
#>  [1] 21.0 21.0 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4 10.4 14.7 32.4 30.4 33.9 21.5 15.5 15.2 13.3 19.2 27.3 26.0 30.4 15.8 19.7 15.0
#> [32] 21.4
#> 
#> $cyl
#>  [1] 6 6 4 6 8 6 8 4 4 6 6 8 8 8 8 8 8 4 4 4 4 8 8 8 8 4 4 4 8 6 8 4
#> 
#> $hp
#>  [1] 110 110  93 110 175 105 245  62  95 123 123 180 180 180 205 215 230  66  52  65  97 150 150 245 175  66  91 113 264 175 335 109
#> 
#> $wt
#>  [1] 2.620 2.875 2.320 3.215 3.440 3.460 3.570 3.190 3.150 3.440 3.440 4.070 3.730 3.780 5.250 5.424 5.345 2.200 1.615 1.835 2.465 3.520 3.435 3.840 3.845 1.935
#> [27] 2.140 1.513 3.170 2.770 3.570 2.780
#> 
#> $vs
#>  [1] 0 0 1 1 0 1 0 1 1 1 1 0 0 0 0 0 0 1 1 1 1 0 0 0 0 1 0 1 0 0 0 1
#> 
#> $am
#>  [1] 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 1 1 1 1 1 1 1
#> 
#> $gear
#>  [1] 4 4 4 3 3 3 3 4 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 4 5 5 5 5 5 4
#> 
#> $carb
#>  [1] 4 4 1 1 2 1 4 2 2 4 4 3 3 3 4 4 4 1 2 1 1 2 2 4 2 1 2 2 4 6 8 2
#> 

# works with any R objects that are internally of type list
mtcars |> pal::list_drop(c("disp", "drat", "qsec"))
#>                      mpg cyl  hp    wt vs am gear carb
#> Mazda RX4           21.0   6 110 2.620  0  1    4    4
#> Mazda RX4 Wag       21.0   6 110 2.875  0  1    4    4
#> Datsun 710          22.8   4  93 2.320  1  1    4    1
#> Hornet 4 Drive      21.4   6 110 3.215  1  0    3    1
#> Hornet Sportabout   18.7   8 175 3.440  0  0    3    2
#> Valiant             18.1   6 105 3.460  1  0    3    1
#> Duster 360          14.3   8 245 3.570  0  0    3    4
#> Merc 240D           24.4   4  62 3.190  1  0    4    2
#> Merc 230            22.8   4  95 3.150  1  0    4    2
#> Merc 280            19.2   6 123 3.440  1  0    4    4
#> Merc 280C           17.8   6 123 3.440  1  0    4    4
#> Merc 450SE          16.4   8 180 4.070  0  0    3    3
#> Merc 450SL          17.3   8 180 3.730  0  0    3    3
#> Merc 450SLC         15.2   8 180 3.780  0  0    3    3
#> Cadillac Fleetwood  10.4   8 205 5.250  0  0    3    4
#> Lincoln Continental 10.4   8 215 5.424  0  0    3    4
#> Chrysler Imperial   14.7   8 230 5.345  0  0    3    4
#> Fiat 128            32.4   4  66 2.200  1  1    4    1
#> Honda Civic         30.4   4  52 1.615  1  1    4    2
#> Toyota Corolla      33.9   4  65 1.835  1  1    4    1
#> Toyota Corona       21.5   4  97 2.465  1  0    3    1
#> Dodge Challenger    15.5   8 150 3.520  0  0    3    2
#> AMC Javelin         15.2   8 150 3.435  0  0    3    2
#> Camaro Z28          13.3   8 245 3.840  0  0    3    4
#> Pontiac Firebird    19.2   8 175 3.845  0  0    3    2
#> Fiat X1-9           27.3   4  66 1.935  1  1    4    1
#> Porsche 914-2       26.0   4  91 2.140  0  1    5    2
#> Lotus Europa        30.4   4 113 1.513  1  1    5    2
#> Ford Pantera L      15.8   8 264 3.170  0  1    5    4
#> Ferrari Dino        19.7   6 175 2.770  0  1    5    6
#> Maserati Bora       15.0   8 335 3.570  0  1    5    8
#> Volvo 142E          21.4   4 109 2.780  1  1    4    2