Parse (R) Markdown as CommonMark XML tree

Parses (R) Markdown file content according to the CommonMark specification and returns it as an XML parse tree.

Usage

md_to_xml(
  md,
  smart_punctuation = FALSE,
  hardbreaks = FALSE,
  normalize = TRUE,
  sourcepos = FALSE,
  extensions = TRUE,
  eol = c("LF", "CRLF", "CR", "LFCR"),
  strip_xml_ns = TRUE
)

Arguments

md

(R) Markdown file content as a character scalar.

smart_punctuation

Whether or not to enable Pandoc's smart extension which converts straight quotes to curly quotes, --- to an em-dash (—), -- to an en-dash (–), and ... to ellipses (…). It also replaces regular spaces after certain abbreviations such as Mr. with non-breaking spaces.

hardbreaks

Whether or not to treat newlines as hard line breaks.

normalize

Consolidate adjacent text nodes.

sourcepos

Include source position attribute in output.

extensions

Enables Github extensions. Can be TRUE (all) FALSE (none) or a character vector with a subset of available extensions.

eol

End of line (EOL) control character sequence. One of

"LF" for the line feed (LF) character ("\n"). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the default.

"CRLF" for the carriage return + line feed (CR+LF) character sequence ("\r\n"). The standard on Microsoft Windows, DOS and some other systems.

"CR" for the carriage return (CR) character ("\r"). The standard on classic Mac OS and some other antiquated systems.

"LFCR" for the line feed + carriage return (LF+CR) character sequence ("\n\r"). The standard on RISC OS and some other exotic systems.

strip_xml_ns

Whether or not to remove the default XML namespace (d1) assigned by commonmark::markdown_xml().

Value

An xml_document.

See also

Examples

pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
  pal::md_to_xml()
#> {xml_document}
#> <document>
#>  [1] <heading level="1">\n  <text xml:space="preserve">INTERNAL</text>\n</heading>
#>  [2] <heading level="2">\n  <text xml:space="preserve">Logo</text>\n</heading>
#>  [3] <paragraph>\n  <text xml:space="preserve">The package logo is based on the vector graphic </text>\n  <link destination="https://thenounproject.com/term/ ...
#>  [4] <paragraph>\n  <text xml:space="preserve">This was fed to </text>\n  <link destination="https://github.com/GuangchuangYu/hexSticker" title="">\n    <tex ...
#>  [5] <code_block info="{r, purl = FALSE}" xml:space="preserve">hexSticker::sticker(subplot = "data-raw/logo/help_yellow.png",\n                    s_x = 1.0, ...
#>  [6] <paragraph>\n  <text xml:space="preserve">Finally, </text>\n  <code xml:space="preserve">data-raw/logo/hexsticker.png</code>\n  <text xml:space="preserv ...
#>  [7] <heading level="2">\n  <text xml:space="preserve">Avoid </text>\n  <code xml:space="preserve">R CMD check</code>\n  <text xml:space="preserve"> notes ab ...
#>  [8] <paragraph>\n  <text xml:space="preserve">Cf. </text>\n  <link destination="https://github.com/tidyverse/magrittr/issues/29#issuecomment-74313262" title ...
#>  [9] <code_block info="{r}" xml:space="preserve">utils::globalVariables(names = c(".",\n                                 # tidyselect fns\n                   ...
#> [10] <heading level="2">\n  <text xml:space="preserve">Constants</text>\n</heading>
#> [11] <code_block info="{r}" xml:space="preserve"># forbidden dots arguments\nforbidden_dots &lt;- list(roxy_tag_value = c("pkgs",\n                           ...
#> [12] <heading level="2">\n  <text xml:space="preserve">Functions</text>\n</heading>
#> [13] <heading level="3">\n  <code xml:space="preserve">env_var_name</code>\n</heading>
#> [14] <paragraph>\n  <text xml:space="preserve">NOTES:</text>\n</paragraph>
#> [15] <list type="bullet" tight="true">\n  <item>\n    <paragraph>\n      <text xml:space="preserve">Only alphanumeric chars plus underscore are </text>\n     ...
#> [16] <code_block info="{r}" xml:space="preserve">env_var_name &lt;- function(...) {\n  \n  as_string(...,\n            sep = "_") %&gt;%\n    stringr::str_re ...
#> [17] <heading level="3">\n  <code xml:space="preserve">get_pkg_config</code>\n</heading>
#> [18] <code_block info="{r}" xml:space="preserve">get_pkg_config &lt;- function(pkg) {\n  \n  checkmate::assert_string(pkg)\n  if (!exists_in_namespace(x = "p ...
#> [19] <heading level="3">\n  <code xml:space="preserve">get_pkg_config_val</code>\n</heading>
#> [20] <code_block info="{r}" xml:space="preserve">get_pkg_config_val &lt;- function(key,\n                               pkg,\n                                ...
#> ...