Get package configuration value

Retrieves a package configuration value in a canonical way. The following configuration sources are consulted in descending order and the first hit is returned:

The R option <pkg>.<key>.

The environment variable <PKG>_<KEY>.

The ad-hoc default value specified via this function's default argument (NULL means unspecified).

The configuration's global default value as specified in the package's configuration metadata (<pkg>::pkg_config$default_value). If no default value is specified (NULL), an error is thrown.

Usage

pkg_config_val(key, pkg, default = NULL)

Arguments

key

Configuration key name. A character scalar.

pkg

Package name. A character scalar.

default

Default value to fall back to if neither the R option <pkg>.<key> nor the environment variable <PKG>_<KEY> is set. If NULL, the default value for key in <pkg>::pkg_config will be used (if defined).

Value

An R object.

Details

This function is intended to be used by package authors who want to expose their package configuration options in a canonical way (as outlined above). For pkg_config_val() to properly work, the configuration metadata must be available in the package's namespace as object pkg_config, which must be a dataframe or tibble with at minimum the columns key (of type character holding the configuration key names) and default_value (of type list holding the default configuration values, if any).

See also

Examples

try(
  pal::pkg_config_val(key = "gen_pkgdown_ref",
                      pkg = "pkgpurl")
)
#> Error in pal::pkg_config_val(key = "gen_pkgdown_ref", pkg = "pkgpurl") : 
#>   Please set the pkgpurl package configuration option gen_pkgdown_ref by either setting the R option pkgpurl.gen_pkgdown_ref or the environment variable
#> `PKGPURL_GEN_PKGDOWN_REF`.