List a function's default parameter values in prose-style

Extracts the default value(s) of a function's definition and returns it in prose style listing.

Usage

prose_ls_fn_param(
  param,
  fn = sys.function(sys.parent()),
  env = parent.frame(),
  as_scalar = TRUE,
  wrap = "`",
  sep = ",",
  last_sep = " or "
)

Arguments

param

Parameter name. A character scalar.

fn

A function or a function name (searched for in env). See base::formals() for details.

env

Environment fn is defined in. See base::formals() for details.

as_scalar

Whether to return the result as a single string concatenated by sep and last_sep.

wrap

String (usually a single character) in which params default values are to be wrapped.

sep

Separator to delimit params default values. Only relevant if as_scalar = TRUE.

last_sep

Separator to delimit the second-last and last one of params default values. Only relevant if as_scalar = TRUE.

Value

A character vector. Of length 1 if as_scalar = TRUE.

Details

This function can be very convenient to avoid duplication in roxygen2 documentation by leveraging inline R code evaluation as follows:

#' @param some_param Some parameter. One of `r prose_ls_fn_param(param = "some_param", fn = "some_fn")`.
some_fn <- function(some_param = c("a", "b", "c")) {
  some_param <- rlang::arg_match(some_param)
  ...
}

Or to list the possible parameter values formatted as an unnumbered list instead, use the inline code `r prose_ls_fn_param(param = "some_param", fn = "some_fn", as_scalar = FALSE) |> as_md_list()` in the example above.

Caveats

base::deparse1() is used internally to get a character representation of non-character default values. Therefore all of deparse()'s fuzziness also applies to this function.

Examples

pal::prose_ls_fn_param(param = ".name_repair",
                       fn = tibble::as_tibble) |>
pal::cat_lines()
#> `"check_unique"`,`"unique"`,`"universal"` or `"minimal"`

pal::prose_ls_fn_param(param = ".name_repair",
                       fn = tibble::as_tibble,
                       as_scalar = FALSE) |>
pal::cat_lines()
#> `"check_unique"`
#> `"unique"`
#> `"universal"`
#> `"minimal"`