Get roxygen2 blocks

Parses the roxygen2 package documentation of a specific R package or from a single .R source code file.

roxy_blocks(pkg = NULL, ..., text = NULL, quiet = TRUE)

Arguments

pkg, text
Either a package name or a character vector of R source code lines to extract the object's roxygen2 tag value from.

...
Further arguments passed on to download.packages(), excluding pkgs, destdir, available and type. Only relevant if pkg is providedDynamic dots are supported.)`

quiet
Whether or not to suppress printing status output from internal processing.

Value

A list of roxy_block objects.

See also

Examples

# Either provide an R source file as a character vector `text` ...
text <- readr::read_lines(paste0("https://raw.githubusercontent.com/r-lib/rlang/",
                                 "db52a58d505b65f58ba922d4752b5b0061f2a98c/R/fn.R"))

pal::roxy_blocks(text = text) |> head(n = 3L)
#> [[1]]
#> <roxy_block> [<text>:39]
#>   $tag
#>     [line:  1] @title 'Create a function' {parsed}
#>     [line:  3] @description '...' {parsed}
#>     [line: 10] @param 'args A named list or pairlist of default argume...' {parsed}
#>     [line: 14] @param 'body A language object representing the code in...' {parsed}
#>     [line: 17] @param 'env The parent environment of the function, def...' {parsed}
#>     [line: 19] @export '' {parsed}
#>     [line: 20] @examples '...' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {parsed}
#>     [????:???] @backref '<text>' {parsed}
#>   $call   new_function <- function(args, body, env = caller_env()) { ...
#>   $object <function> 
#>     $topic new_function
#>     $alias new_function
#> 
#> [[2]]
#> <roxy_block> [<text>:50]
#>   $tag
#>     [line:  1] @title 'Name of a primitive function' {parsed}
#>     [line: 47] @param 'prim A primitive function such as [base::c()].' {parsed}
#>     [line: 48] @keywords 'internal' {parsed}
#>     [line: 49] @export '' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {parsed}
#>     [????:???] @backref '<text>' {parsed}
#>   $call   prim_name <- function(prim) { ...
#>   $object <function> 
#>     $topic prim_name
#>     $alias prim_name
#> 
#> [[3]]
#> <roxy_block> [<text>:91]
#>   $tag
#>     [line:  1] @title 'Extract arguments from a function' {parsed}
#>     [line:  3] @description '`fn_fmls()` returns a named list of formal argu...' {parsed}
#>     [line:  9] @details 'Unlike `formals()`, these helpers throw an erro...' {parsed}
#>     [line: 73] @param 'fn A function. It is lookep up in the calling f...' {parsed}
#>     [line: 75] @seealso '[call_args()] and [call_args_names()]' {parsed}
#>     [line: 76] @export '' {parsed}
#>     [line: 77] @examples '...' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {parsed}
#>     [????:???] @backref '<text>' {parsed}
#>   $call   fn_fmls <- function(fn = caller_fn()) { ...
#>   $object <function> 
#>     $topic fn_fmls
#>     $alias fn_fmls
#> 

# ... or provide a package name
try(
  pal::roxy_blocks(pkg = "rlang",
                   repos = "https://cloud.r-project.org") |>
    head(n = 3L)
)
#> Warning: [/home/salim/.cache/pal-roxy_tag_value-rlang/rlang/R/attr.R:249] @description mismatched braces or quotes
#> Warning: [/home/salim/.cache/pal-roxy_tag_value-rlang/rlang/R/call.R:322] @section mismatched braces or quotes
#> Warning: [/home/salim/.cache/pal-roxy_tag_value-rlang/rlang/R/nse-defuse.R:3] @description mismatched braces or quotes
#> Warning: [/home/salim/.cache/pal-roxy_tag_value-rlang/rlang/R/nse-force.R:66] @section mismatched braces or quotes
#> Warning: [/home/salim/.cache/pal-roxy_tag_value-rlang/rlang/R/parse.R:3] @description mismatched braces or quotes
#> * installing *source* package ‘rlang’ ...
#> ** package ‘rlang’ successfully unpacked and MD5 sums checked
#> ** using staged installation
#> ** libs
#> gcc -std=gnu99 -std=gnu11 -I"/usr/share/R/include" -DNDEBUG -I./lib/    -fvisibility=hidden -fpic  -g -O2 -fdebug-prefix-map=/build/r-base-i2PIHO/r-base-4.1.2=. -fstack-protector-strong -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 -g  -UNDEBUG -Wall -pedantic -g -O0 -fdiagnostics-color=always -c capture.c -o capture.o
#> gcc -std=gnu99 -std=gnu11 -I"/usr/share/R/include" -DNDEBUG -I./lib/    -fvisibility=hidden -fpic  -g -O2 -fdebug-prefix-map=/build/r-base-i2PIHO/r-base-4.1.2=. -fstack-protector-strong -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 -g  -UNDEBUG -Wall -pedantic -g -O0 -fdiagnostics-color=always -c export.c -o export.o
#> gcc -std=gnu99 -std=gnu11 -I"/usr/share/R/include" -DNDEBUG -I./lib/    -fvisibility=hidden -fpic  -g -O2 -fdebug-prefix-map=/build/r-base-i2PIHO/r-base-4.1.2=. -fstack-protector-strong -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 -g  -UNDEBUG -Wall -pedantic -g -O0 -fdiagnostics-color=always -c internal.c -o internal.o
#> gcc -std=gnu99 -std=gnu11 -I"/usr/share/R/include" -DNDEBUG -I./lib/    -fvisibility=hidden -fpic  -g -O2 -fdebug-prefix-map=/build/r-base-i2PIHO/r-base-4.1.2=. -fstack-protector-strong -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 -g  -UNDEBUG -Wall -pedantic -g -O0 -fdiagnostics-color=always -c lib.c -o lib.o
#> gcc -std=gnu99 -std=gnu11 -I"/usr/share/R/include" -DNDEBUG -I./lib/    -fvisibility=hidden -fpic  -g -O2 -fdebug-prefix-map=/build/r-base-i2PIHO/r-base-4.1.2=. -fstack-protector-strong -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 -g  -UNDEBUG -Wall -pedantic -g -O0 -fdiagnostics-color=always -c version.c -o version.o
#> gcc -std=gnu99 -std=gnu11 -shared -L/usr/lib/R/lib -Wl,-Bsymbolic-functions -Wl,-z,relro -o rlang.so capture.o export.o internal.o lib.o version.o -L/usr/lib/R/lib -lR
#> installing to /tmp/Rtmpava2GC/devtools_install_20dc31ff0b85c/00LOCK-rlang/00new/rlang/libs
#> ** checking absolute paths in shared objects and dynamic libraries
#> * DONE (rlang)
#> [[1]]
#> <roxy_block> [arg.R:26]
#>   $tag
#>     [line:  1] @title 'Match an argument to a character vector' {parsed}
#>     [line:  3] @description '...' {parsed}
#>     [line: 15] @param 'arg A symbol referring to an argument accepting...' {parsed}
#>     [line: 16] @return 'The string supplied to `arg`.' {parsed}
#>     [line: 17] @importFrom 'utils adist' {parsed}
#>     [line: 18] @export '' {parsed}
#>     [line: 19] @examples '...' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {parsed}
#>     [????:???] @backref '/home/salim/.cache/pal-roxy_tag_value-rlang/rla...' {parsed}
#>   $call   arg_match <- function(arg, values = NULL) { ...
#>   $object <function> 
#>     $topic arg_match
#>     $alias arg_match
#> 
#> [[2]]
#> <roxy_block> [arg.R:73]
#>   $tag
#>     [line: 50] @description '...' {parsed}
#>     [line: 57] @param 'values The possible values that `arg` can take.' {parsed}
#>     [line: 58] @param 'arg_nm The label to be used for `arg` in error ...' {parsed}
#>     [line: 59] @rdname 'arg_match' {parsed}
#>     [line: 60] @export '' {parsed}
#>     [line: 61] @examples '...' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {parsed}
#>     [????:???] @backref '/home/salim/.cache/pal-roxy_tag_value-rlang/rla...' {parsed}
#>   $call   arg_match0 <- function(arg, values, arg_nm = as_label(substitute(arg))) { ...
#>   $object <function> 
#>     $topic arg_match0
#>     $alias arg_match0
#> 
#> [[3]]
#> <roxy_block> [arg.R:237]
#>   $tag
#>     [line:  1] @title 'Generate or handle a missing argument' {parsed}
#>     [line:130] @description '...' {parsed}
#>     [line:146] @section 'Other ways to reify the missing argument:...' {parsed}
#>     [line:157] @section 'Fragility of the missing argument object:...' {parsed}
#>     [line:183] @section 'Life cycle:...' {parsed}
#>     [line:188] @param 'x An object that might be the missing argument.' {parsed}
#>     [line:189] @export '' {parsed}
#>     [line:190] @examples '...' {parsed}
#>     [????:???] @usage '<generated>' {parsed}
#>     [????:???] @.formals '<generated>' {unparsed}
#>     [????:???] @backref '/home/salim/.cache/pal-roxy_tag_value-rlang/rla...' {parsed}
#>   $call   missing_arg <- function() { ...
#>   $object <function> 
#>     $topic missing_arg
#>     $alias missing_arg
#>