Convert to sentence case with trailing punctuation mark

Converts the input to a character vector and ensures it starts with an upper case letter and ends with the specified punctuation mark.

sentenceify(x, punctuation_mark = ".")

Arguments

x
Input to be converted to sentence case, typically a character vector.

punctuation_mark
Punctuation mark to be appended to x. A character scalar.

Value

A character vector.

Details

Note that this function doesn't alter any characters in x other than the first and the last one.

See also

Examples

pal::sentenceify("no verb, no sentence")
#> [1] "No verb, no sentence."

# punctuation mark won't be duplicated if already existing
pal::sentenceify(c("I've made my point",
                   "good point."))
#> [1] "I've made my point." "Good point."