Strip YAML header from R Markdown

Extracts the body from R Markdown file content, stripping a possible YAML metadata block at the beginning.

strip_yaml_header(rmd, eol = c("LF", "CRLF", "CR", "LFCR"))

Arguments

rmd
R Markdown file content as a character scalar.

eol
End of line (EOL) control character sequence. One of

"LF" for the line feed (LF) character ("\n"). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the default.

"CRLF" for the carriage return + line feed (CR+LF) character sequence ("\r\n"). The standard on Microsoft Windows, DOS and some other systems.

"CR" for the carriage return (CR) character ("\r"). The standard on classic Mac OS and some other antiquated systems.

"LFCR" for the line feed + carriage return (LF+CR) character sequence ("\n\r"). The standard on RISC OS and some other exotic systems.

Value

Body of the R Markdown file as a character vector of lines.

Details

Note that for the R Markdown file format, the YAML metadata block must occur at the beginning of the document (and there can be only one). Additional whitespace characters (incl. newlines) before the YAML metadata block are allowed.

See also

Examples

library(magrittr)

pal::gh_text_file(path = "README.Rmd",
                  owner = "salim-b",
                  name = "pal") %T>%
  pal::strip_yaml_header()
#> [1] "---\noutput: pal::gitlab_document\n---\n\n# `r paste0(pal::desc_value(\"Package\"), \": \", pal::desc_value(\"Title\"))` <a href=\"https://rpkg.dev/pal/\"><img src=\"man/figures/logo.png\" alt=\"Package Logo\" class=\"pkg-logo\" align=\"right\" height=\"180\"/></a>\n\n`r pal::desc_value(\"Description\")`[^1]\n\nThis is arguably a [*low-cohesion*](https://en.wikipedia.org/wiki/Cohesion_(computer_science)) package.\n\n[^1]: If you discover one or more of the features of this package in a more popular tidyverse-style package, feel free [to report\nit](https://gitlab.com/salim_b/r/pkgs/pal/-/issues/new) to me and I will consider deprecating it. I strongly support standardization efforts and the avoidance\nof unnecessary duplication.\n\n```{r, pkgdown, eval = !isTRUE(getOption(\"pal.build_readme.is_pkgdown\")), results = \"asis\", echo = FALSE}\npkgsnip::md_snip(name = \"pkgdown_notice\") %>%\n  paste0(\"## Documentation\\n\\n\", .) %>%\n  pal::cat_lines()\n```\n\n## Installation\n\n```{r, install-snippet-dev, child = pkgsnip::snip_path(\"installation-notice_dev-version_gitlab.Rmd\")}\n```\n\n## Development\n\n### R Markdown format\n\n```{r, rmd-pkg-snippet, child = pkgsnip::snip_path(\"rmd-package-notice.Rmd\")}\n```\n\n### Coding style\n\n```{r, coding-style-snippet, child = pkgsnip::snip_path(\"coding-style-notice.Rmd\")}\n```\n\n## See also\n\n-   The [**xfun**](https://yihui.org/xfun/) package by Yihui Xie, a non-overlapping collection of miscellaneous R functions.\n"